Network message queue functions.
These functions provide a simple interface for managing a queue of
network messages. An underlaying structure is implemented as a linked
list which allows for fast pushes (enqueing), though an array (buffer)
is used as a medium for transfer, which contains serialized messages.

This buffer is nul terminated and can be passed to send(3).