Do daemonization properly.
If daemonization was done after opening an interface with libpcap, no
data were captured and error notification message was sent immediatelly
after.

In this patch, parent process opens a pipe(2) and attempts to read from
it in blocking mode, while child process does all the necessary magic.
If no error is encountered, child process closes writable end of the
pipe and blocks indefinitelly in the main loop; otherwise an error
message is sent to the pipe where is received by the parent process and
printed to stderr.