Obtain a network mask of a interface.
If we want to use bpf's broadcast directive pcap_compile requires a
netmask to be provided. In case we are not capturing in monitor mode,
use pcap_lookupnet to obtain the netmask and pass it to pcap_compile.

If pcap_lookupnet fails, ignore the error and use PCAP_NETMASK_UNKNOWN.
If broadcast directive is applied with PCAP_NETMASK_UNKNOWN,
pcap_compile will fail.